MoDora-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "modora-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "python-Levenshtein",
  "tqdm",
  "numpy",
  "opencv-python",
  "openai",
  "pymupdf",
  "Pillow",
  "nvidia-ml-py",
  "huggingface_hub",
  "transformers>=4.45.0",
  "pandas",
  "seaborn",
  "matplotlib",
  "python-multipart",
  "lmdeploy==0.12.0",
  "paddleocr",
  "paddlex>=3.4.2",
  "chromadb",
]

[project.scripts]
modora = "modora.lab.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
